BRE-178: Support individual tags for components --- Users may need to specify a specific version for a single component.
Argo CD Image updater will need to update these to digests when a new image is pushed to the registry

This adds the ability to define tags for each Bitwarden component and the MSSQL migrator utility. Allowing us to use a tag for each component if we wish to do so.

Deployments now have a $tag variable and this value is set on the logic of
Check for .tag value > Check for override value > Set value from Default

There was also some refactor for Key Connector to bring it in alignment with this new pattern, we no longer use `version` but instead use `tag` for key connector.